pyproject.toml
[tool.poetry]
name = "tea_cloud"
version = "0.1.0"
description = "A Discord bot to generate a word cloud of the past day's discussion."
authors = ["Émilia Decaudin"]
license = "Copyright (c) Émilia Decaudin"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
wordcloud = "^1.9.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pre-commit = "^3.7.1"
ruff = "^0.4.9"

[tool.mypy]
disallow_untyped_defs = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
check_untyped_defs = true
files = ["tea_cloud"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["A", "B", "C4", "E", "F", "I", "N", "RUF", "SLF", "UP", "W"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
